+++ controllers/flux/volumes.go
@@ -13,16 +13,11 @@ package controllers
 import (
 	"context"
 	"fmt"
-	"path"
 
 	api "github.com/flux-framework/flux-operator/api/v1alpha2"
 
 	corev1 "k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/api/errors"
-	"k8s.io/apimachinery/pkg/api/resource"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/types"
-	ctrl "sigs.k8s.io/controller-runtime"
 )
 
 // Shared function to return consistent set of volume mounts
@@ -120,25 +115,11 @@ func getVolumes(cluster *api.MiniCluster) []corev1.Volume {
 	// This can be a claim, secret, or config map
 	existingVolumes := getExistingVolumes(cluster.ExistingContainerVolumes())
 	volumes = append(volumes, existingVolumes...)
-
-	// Add claims for storage types requested
-	// These are created by the Flux Operator and should be Container Storage Interfaces (CSI)
-	for volumeName := range cluster.Spec.Volumes {
-		newVolume := corev1.Volume{
-			Name: volumeName,
-			VolumeSource: corev1.VolumeSource{
-				PersistentVolumeClaim: &corev1.PersistentVolumeClaimVolumeSource{
-					ClaimName: fmt.Sprintf("%s-claim", volumeName),
-				},
-			},
-		}
-		volumes = append(volumes, newVolume)
-	}
 	return volumes
 }
 
 // Get Existing volumes for the MiniCluster
-func getExistingVolumes(existing map[string]api.MiniClusterExistingVolume) []corev1.Volume {
+func getExistingVolumes(existing map[string]api.ContainerVolume) []corev1.Volume {
 	volumes := []corev1.Volume{}
 	for volumeName, volumeMeta := range existing {
 
@@ -195,86 +176,6 @@ func getExistingVolumes(existing map[string]api.MiniClusterExistingVolume) []cor
 	return volumes
 }
 
-// createPersistentVolume creates a volume in /mnt
-func CreatePersistentVolume(
-	cluster *api.MiniCluster,
-	volumeName string,
-	volume api.MiniClusterVolume,
-) *corev1.PersistentVolume {
-
-	// We either support a hostpath (miniKube) or a Container Storage Interface (CSI)
-	var pvsource corev1.PersistentVolumeSource
-	if volume.StorageClass == "hostpath" {
-
-		pvsource = corev1.PersistentVolumeSource{
-			HostPath: &corev1.HostPathVolumeSource{
-				Path: path.Join(volume.Path),
-			},
-		}
-
-	} else {
-
-		// VolumeHandle defaults to storage class name
-		// unless it is explicitly different!
-		volumeHandle := volume.StorageClass
-		if volume.VolumeHandle != "" {
-			volumeHandle = volume.VolumeHandle
-		}
-		pvsource = corev1.PersistentVolumeSource{
-			CSI: &corev1.CSIPersistentVolumeSource{
-
-				// Choose for the user for now.
-				Driver: volume.Driver,
-
-				// Name in storageclass metadata, also what we use for name
-				VolumeHandle: volumeHandle,
-				NodePublishSecretRef: &corev1.SecretReference{
-					Namespace: volume.SecretNamespace,
-					Name:      volume.Secret,
-				},
-				ControllerPublishSecretRef: &corev1.SecretReference{
-					Namespace: volume.SecretNamespace,
-					Name:      volume.Secret,
-				},
-				NodeStageSecretRef: &corev1.SecretReference{
-					Namespace: volume.SecretNamespace,
-					Name:      volume.Secret,
-				},
-				VolumeAttributes: volume.Attributes,
-			},
-		}
-	}
-
-	newVolume := &corev1.PersistentVolume{
-		TypeMeta: metav1.TypeMeta{
-			Kind:       "PersistentVolume",
-			APIVersion: "v1",
-		},
-		ObjectMeta: metav1.ObjectMeta{
-			Name:        volumeName,
-			Namespace:   cluster.Namespace,
-			Annotations: volume.Annotations,
-			Labels:      volume.Labels,
-		},
-
-		Spec: corev1.PersistentVolumeSpec{
-			PersistentVolumeReclaimPolicy: corev1.PersistentVolumeReclaimDelete,
-			AccessModes:                   []corev1.PersistentVolumeAccessMode{corev1.ReadWriteMany},
-
-			// This is a path in the minikube vm or on the node
-			PersistentVolumeSource: pvsource,
-			StorageClassName:       volume.StorageClass,
-		},
-	}
-	// Capacity is optional for some storage like Google Cloud
-	if volume.Capacity != "" {
-		newVolume.Spec.Capacity = map[corev1.ResourceName]resource.Quantity{
-			corev1.ResourceStorage: resource.MustParse(volume.Capacity),
-		}
-	}
-	return newVolume
-}
-
 func (r *MiniClusterReconciler) getExistingPersistentVolume(
 	ctx context.Context,
 	cluster *api.MiniCluster,
@@ -293,63 +194,6 @@ func (r *MiniClusterReconciler) getExistingPersistentVolume(
 	return existing, err
 }
 
-// getPersistentVolume creates the PV for the curve certificate (to be written once)
-func (r *MiniClusterReconciler) getPersistentVolume(
-	ctx context.Context,
-	cluster *api.MiniCluster,
-	volumeName string,
-	volume api.MiniClusterVolume) (*corev1.PersistentVolume, ctrl.Result, error,
-) {
-
-	existing, err := r.getExistingPersistentVolume(ctx, cluster, volumeName)
-
-	if err != nil {
-
-		// Case 1: not found yet, and hostfile is ready (recreate)
-		if errors.IsNotFound(err) {
-
-			// The volume "<name>-volume" will be under /mnt/<name>
-			volume := CreatePersistentVolume(cluster, volumeName, volume)
-			r.log.Info(
-				"✨ Creating MiniCluster Mounted Volume ✨",
-				"Type", volumeName,
-				"Namespace", cluster.Namespace,
-				"Name", volumeName,
-			)
-
-			err = r.New(ctx, volume)
-			if err != nil {
-				r.log.Error(
-					err, "Failed to create MiniCluster Mounted Volume",
-					"Type", volumeName,
-					"Namespace", cluster.Namespace,
-					"Name", volumeName,
-				)
-				return existing, ctrl.Result{}, err
-			}
-			// Successful - return and requeue
-			return volume, ctrl.Result{Requeue: true}, nil
-
-		} else if err != nil {
-			r.log.Error(err, "Failed to get MiniCluster Mounted Volume")
-			return existing, ctrl.Result{}, err
-		}
-
-	} else {
-
-		r.log.Info(
-			"🎉 Found existing MiniCluster Mounted Volume",
-			"Type", volumeName,
-			"Namespace", existing.Namespace,
-			"Name", existing.Name,
-		)
-	}
-
-	// Always set owner to controller, whether created or found
-	ctrl.SetControllerReference(cluster, existing, r.Scheme)
-	return existing, ctrl.Result{}, err
-}
-
 // getExistingVolumeClaim gets an existing volume claim
 func (r *MiniClusterReconciler) getExistingPersistentVolumeClaim(
 	ctx context.Context,
@@ -366,95 +210,4 @@ func (r *MiniClusterReconciler) getExistingPersistentVolumeClaim(
 		existing,
 	)
 	return existing, err
-}
-
-// getPersistentVolume creates the PVC claim for the curve certificate (to be written once)
-func (r *MiniClusterReconciler) getPersistentVolumeClaim(
-	ctx context.Context,
-	cluster *api.MiniCluster,
-	volumeName string,
-	volume api.MiniClusterVolume,
-) (*corev1.PersistentVolumeClaim, ctrl.Result, error) {
-
-	claimName := fmt.Sprintf("%s-claim", volumeName)
-	existing, err := r.getExistingPersistentVolumeClaim(ctx, cluster, claimName)
-
-	if err != nil {
-
-		// Case 1: not found yet, and hostfile is ready (recreate)
-		if errors.IsNotFound(err) {
-			volume := CreatePersistentVolumeClaim(cluster, claimName, volume)
-			ctrl.SetControllerReference(cluster, volume, r.Scheme)
-			r.log.Info(
-				"✨ Creating MiniCluster Mounted Volume ✨",
-				"Type", claimName,
-				"Namespace", cluster.Namespace,
-				"Name", claimName,
-			)
-			err = r.New(ctx, volume)
-
-			// This is a creation error we need to report back
-			if err != nil {
-				r.log.Error(
-					err, "❌ Failed to create MiniCluster Mounted Volume",
-					"Type", claimName,
-					"Namespace", volume.Namespace,
-					"Name", claimName,
-				)
-				return existing, ctrl.Result{}, err
-			}
-			// Successful - return and requeue
-			return volume, ctrl.Result{Requeue: true}, nil
-
-		} else if err != nil {
-			r.log.Error(err, "Failed to get MiniCluster Mounted Volume")
-			return existing, ctrl.Result{}, err
-		}
-
-	} else {
-
-		r.log.Info("🎉 Found existing MiniCluster Mounted Volume Claim",
-			"Type", claimName,
-			"Namespace", existing.Namespace,
-			"Name", existing.Name,
-		)
-	}
-	return existing, ctrl.Result{}, err
-}
-
-// createPersistentVolumeClaim generates a PVC
-func CreatePersistentVolumeClaim(
-	cluster *api.MiniCluster,
-	volumeName string,
-	volume api.MiniClusterVolume,
-) *corev1.PersistentVolumeClaim {
-
-	volumeMode := corev1.PersistentVolumeFilesystem
-
-	// Create a new RWX persistent volume claim
-	newVolume := &corev1.PersistentVolumeClaim{
-		TypeMeta: metav1.TypeMeta{
-			Kind:       "PersistentVolumeClaim",
-			APIVersion: "v1",
-		},
-		ObjectMeta: metav1.ObjectMeta{
-			Name:        volumeName,
-			Namespace:   cluster.Namespace,
-			Annotations: volume.ClaimAnnotations,
-		},
-
-		Spec: corev1.PersistentVolumeClaimSpec{
-			StorageClassName: &volume.StorageClass,
-			AccessModes: []corev1.PersistentVolumeAccessMode{
-				corev1.ReadWriteMany,
-			},
-			VolumeMode: &volumeMode,
-			Resources: corev1.ResourceRequirements{
-				Requests: corev1.ResourceList{
-					corev1.ResourceStorage: resource.MustParse(volume.Capacity),
-				},
-			},
-		},
-	}
-	return newVolume
 }