+++ examples/tests/volumes/post-run.sh
@@ -1,4 +1,0 @@
-#!/bin/bash
-
-echo "Cleaning up /tmp/data in minikube"
-minikube ssh -- sudo rm -rf /tmp/data